circles appear, but very low performance, works only when predators are around 5

+++ src/app/config.java
@@ -2,7 +2,7 @@ package app;
 
 public class config {
 
-    public static final double window_width = 1000;
-    public static final double window_height = 700;
+    public static final int window_width = 1000;
+    public static final int window_height = 700;
 
 }
